+++ buttons/helpers/buttons_merged_menu.js
@@ -1,5 +1,5 @@
 ﻿'use strict';
-//27/04/26
+//12/06/26
 
 /* exported createButtonsMenu, onRbtnUpImportSettings */
 
@@ -893,7 +893,7 @@ function createButtonsMenu(name) {
 				barProperties.bAutoUpdateCheck[1] = !barProperties.bAutoUpdateCheck[1];
 				overwriteProperties(barProperties);
 				if (barProperties.bAutoUpdateCheck[1]) {
-					if (typeof checkUpdate === 'undefined') { include('..\\..\\helpers\\helpers_xxx_web_update.js'); }
+					if (typeof checkUpdate === 'undefined') { include('..\\..\\helpers\\helpers_xxx_web.js'); }
 					checkUpdate({ bDownload: globSettings.bAutoUpdateDownload, bOpenWeb: globSettings.bAutoUpdateOpenWeb, bDisableWarning: false });
 				}
 			}
@@ -902,7 +902,7 @@ function createButtonsMenu(name) {
 		menu.newSeparator(subMenu);
 		menu.newEntry({
 			menuName: subMenu, entryText: 'Check for updates...', func: () => {
-				if (typeof checkUpdate === 'undefined') { include('..\\..\\helpers\\helpers_xxx_web_update.js'); }
+				if (typeof checkUpdate === 'undefined') { include('..\\..\\helpers\\helpers_xxx_web.js'); }
 				checkUpdate({ bDownload: globSettings.bAutoUpdateDownload, bOpenWeb: globSettings.bAutoUpdateOpenWeb, bDisableWarning: false })
 					.then((result) => {
 						if (!result) { fb.ShowPopupMessage('No updates found.', window.FullPanelName + ': Update check'); }
@@ -1050,7 +1050,7 @@ function onRbtnUpImportSettings(properties = this.properties || {}) {
 				properties.bAutoUpdateCheck[1] = !properties.bAutoUpdateCheck[1];
 				overwriteProperties(properties);
 				if (properties.bAutoUpdateCheck[1]) {
-					if (typeof checkUpdate === 'undefined') { include('..\\..\\helpers\\helpers_xxx_web_update.js'); }
+					if (typeof checkUpdate === 'undefined') { include('..\\..\\helpers\\helpers_xxx_web.js'); }
 					checkUpdate({ bDownload: globSettings.bAutoUpdateDownload, bOpenWeb: globSettings.bAutoUpdateOpenWeb, bDisableWarning: false });
 				}
 			}
@@ -1059,7 +1059,7 @@ function onRbtnUpImportSettings(properties = this.properties || {}) {
 		menu.newSeparator(subMenu);
 		menu.newEntry({
 			menuName: subMenu, entryText: 'Check for updates...', func: () => {
-				if (typeof checkUpdate === 'undefined') { include('..\\..\\helpers\\helpers_xxx_web_update.js'); }
+				if (typeof checkUpdate === 'undefined') { include('..\\..\\helpers\\helpers_xxx_web.js'); }
 				checkUpdate({ bDownload: globSettings.bAutoUpdateDownload, bOpenWeb: globSettings.bAutoUpdateOpenWeb, bDisableWarning: false })
 					.then((result) => {
 						if (!result) { fb.ShowPopupMessage('No updates found.', window.FullPanelName + ': Update check'); }